Fix: potential out of memory issue with large collections
Trying to get the size of a collection will cause a SELECT * query to be run, and since buffered queries are in use by default, PHP can run out of memory when selecting all from large tables. This change has the pager run a Count aggregate instead of asking for the size of the unpaged collection.